feat(arc-1522): add remaining redirects from the redirect plan
https://meemoo.atlassian.net/browse/ARC-1522
