+++ next-i18next.config.js
@@ -4,21 +4,30 @@ const I18NextHttpBackend = require('i18next-http-backend');
 const _ = require('lodash');
 
 module.exports = {
-	i18n: {
-		defaultLocale: 'nl',
-		locales: ['nl'],
-		backend: {
-			loadPath: `${process.env.PROXY_URL}/translations/nl.json`,
-		},
+	getI18n: (proxyUrl) => {
+		return {
+			defaultLocale: 'nl',
+			locales: ['nl'],
+			backend: {
+				loadPath: `${proxyUrl}/translations/nl.json`,
+			},
+			i18n: {
+				defaultLocale: 'nl',
+				locales: ['nl'],
+				backend: {
+					loadPath: `${proxyUrl}/translations/nl.json`,
+				},
+			},
+			use: [I18NextHttpBackend],
+			ns: ['common', 'admin_core'],
+			serializeConfig: false,
+			parseMissingKeyHandler: (key) => {
+				if (key.includes('___')) {
+					return `${_.upperFirst(_.lowerCase(key.split('___').pop()))} ***`;
+				}
+				return `${key} ***`;
+			},
+			debug: false,
+		};
 	},
-	use: [I18NextHttpBackend],
-	ns: ['common', 'admin_core'],
-	serializeConfig: false,
-	parseMissingKeyHandler: (key) => {
-		if (key.includes('___')) {
-			return `${_.upperFirst(_.lowerCase(key.split('___').pop()))} ***`;
-		}
-		return `${key} ***`;
-	},
-	debug: false,
 };
