--- src/modules/i18n/wrappers/with-i18n/index.ts
@@ -1,1 +1,0 @@
-export * from './with-i18n';
