--- src/modules/i18n/wrappers/with-i18n/with-i18n.types.ts
@@ -1,3 +1,0 @@
-import type { SSRConfig } from 'next-i18next';
-
-export type WithI18n = SSRConfig;
