+++ src/modules/ie-objects/hooks/get-ie-objects-info.ts
@@ -5,8 +5,16 @@ import { QUERY_KEYS } from '@shared/const/query-keys';
 import { IeObjectsService } from './../services';
 import { IeObject } from './../types';
 
-export const useGetIeObjectsInfo = (id: string): UseQueryResult<IeObject> => {
-	return useQuery([QUERY_KEYS.getIeObjectsInfo, { id }], () => IeObjectsService.getById(id), {
-		keepPreviousData: false,
-	});
+export const useGetIeObjectsInfo = (
+	id: string,
+	options: {
+		keepPreviousData: boolean;
+		enabled: boolean;
+	} = { keepPreviousData: true, enabled: true }
+): UseQueryResult<IeObject> => {
+	return useQuery(
+		[QUERY_KEYS.getIeObjectsInfo, { id }],
+		() => IeObjectsService.getById(id),
+		options
+	);
 };
