+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { appWithTranslation, UserConfig } from 'next-i18next';
+import { appWithTranslation } from 'next-i18next';
 import { AppProps } from 'next/app';
+import getConfig from 'next/config';
 import { ReactElement } from 'react';
 
 import { AppLayout } from '@shared/layouts/AppLayout';
@@ -9,8 +10,10 @@ import { wrapper } from '@shared/store';
 
 import 'styles/main.scss';
 
-import NextI18nextConfig from '../../next-i18next.config';
+import { getI18n } from '../../next-i18next.config';
 import pkg from '../../package.json';
+
+const { publicRuntimeConfig } = getConfig();
 
 const queryClient = new QueryClient({
 	defaultOptions: {
@@ -38,4 +41,6 @@ function MyApp({ Component, pageProps }: AppProps): ReactElement {
 	);
 }
 
-export default wrapper.withRedux(appWithTranslation(MyApp, NextI18nextConfig as UserConfig));
+export default wrapper.withRedux(
+	appWithTranslation(MyApp, getI18n(publicRuntimeConfig.PROXY_URL) as any)
+);
